+++ config/routes.rb
@@ -1103,6 +1103,7 @@ Rails.application.routes.draw do
   post 'main/logout', controller: 'main', action: 'logout'
 
   namespace :jupyter do
+    post 'authenticate', controller: 'jupyter_submissions', action: 'create_session'
     post 'submit', controller: 'jupyter_submissions', action: 'submit'
   end